requirements.txt
Pillow==10.2.0
stegano==0.11.1
python-magic-bin==0.4.14
Flask==3.0.2
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5 